refactor namespace restriction fixtures into local conftest
Move fixtures from parent conftest.py into the namespace_restriction
package conftest.py. Extract inline schema dict to a dedicated fixture.
Fix docstring descriptions.
