+++ backend/tests/component/core/schema_manager/conftest.py
@@ -3,7 +3,6 @@ from typing import Any
 import pytest
 
 from infrahub.core.constants import BranchSupportType, InfrahubKind
-from infrahub.core.schema import SchemaRoot
 
 
 def _get_schema_by_kind(full_schema: dict[str, Any], kind: str) -> dict[str, Any]:
@@ -389,47 +388,3 @@ def schema_diff_attr_inheritance_types() -> dict[str, Any]:
         ],
     }
     return FULL_SCHEMA
-
-
-@pytest.fixture
-def correct_schema_generic_with_namespace_restriction() -> SchemaRoot:
-    """One generic with namespace restriction and one inherited node, which does follow the namespace restriction rule"""
-    schema = {
-        "generics": [
-            {
-                "name": "Generic",
-                "namespace": "Animal",
-                "display_labels": ["name__value"],
-                "order_by": ["name__value"],
-                "attributes": [{"name": "name", "kind": "Text"}],
-                "restricted_namespaces": ["Animal"],
-            }
-        ],
-        "nodes": [
-            {
-                "name": "Dog",
-                "namespace": "Animal",
-                "description": "A dog which follows the animal namespace restriction rule",
-                "attributes": [{"name": "dog", "kind": "Text"}],
-                "inherit_from": ["AnimalGeneric"],
-            }
-        ],
-    }
-    return SchemaRoot(**schema)
-
-
-@pytest.fixture
-def incorrect_schema_inherits_from_generic_core_repository() -> SchemaRoot:
-    """One node which inherits from CoreRepository. As the namespace is restricted, this should not be allowed."""
-    schema = {
-        "nodes": [
-            {
-                "name": "ChildrenRepo",
-                "namespace": "Child",
-                "description": "A repository inheriting from the core generic repository",
-                "attributes": [{"name": "children_repository", "kind": "Text"}],
-                "inherit_from": ["CoreGenericRepository"],
-            }
-        ],
-    }
-    return SchemaRoot(**schema)
